adk-backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adk-backend"
version = "0.1.0"
description = "Google ADK 기반 BigQuery 에이전트 백엔드"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Data Platform Team" }
]
dependencies = [
  "google-adk>=1.10.0",
  "google-cloud-bigquery>=3.13.0",
  "google-cloud-bigquery-storage>=2.24.0",
  "google-cloud-aiplatform>=1.65.0",
  "google-cloud-geminidataanalytics>=0.5.0",
  "google-cloud-datacatalog>=3.20.0",
  "pydantic-settings>=2.4.0",
  "fastapi>=0.114.0",
  "uvicorn[standard]>=0.30.0",
  "python-dotenv>=1.0.1",
  # NLP 백엔드 통합을 위한 추가 의존성
  "python-multipart>=0.0.6",
  "starlette>=0.27.0",
  "pydantic>=2.5.0",
  "google-auth>=2.23.4",
  "google-auth-oauthlib>=1.1.0",
  "google-auth-httplib2>=0.1.1",
  "openai>=1.3.7",
  "structlog>=23.2.0",
  "requests>=2.31.0",
  "aiofiles>=23.2.1",
  "typing-extensions>=4.8.0",
  "python-dateutil>=2.8.2",
  "orjson>=3.9.10",
  "regex>=2023.10.3",
  "psutil>=5.9.6",
  "google-genai>=0.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.23.0",
  "httpx>=0.27.0",
  "ruff>=0.5.0",
]

[tool.pytest.ini_options]
addopts = "-ra"
asyncio_mode = "auto"
pythonpath = ["src"]
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow running tests (deselect with '-m \"not slow\"')",
]

[tool.ruff]
line-length = 100
